Add files via upload
Level,java handles reading level files via a scanner which scans for characters which represent different objects that extend from the LevelObject class, and initializes the current level that the player will be playing through by adding all level objects to an arraylist and being able to print the level they are on in ASCII to the terminal. LevelObject is the superclass for all static objects in the level, such as Walls, Doors, Teleporters, as well as empty spaces which are mostly for formatting the ASCII output. Level.java is also able to currently take in the player's X and Y coordinates and display them in the level layout.